retrieval_service/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pytest
pytest-mock
httpx